+++ lib/ductwork/cli.rb
@@ -8,12 +8,12 @@ module Ductwork
       def start!(args)
         options = parse_options(args)
         Ductwork.configuration = Configuration.new(**options)
-        Ductwork.configuration.logger = if Ductwork.configuration.logger_source == "rails"
-                                          Rails.logger
-                                        else
-                                          Ductwork::Configuration::DEFAULT_LOGGER
-                                        end
-        Ductwork.configuration.logger.level = Ductwork.configuration.logger_level
+        Ductwork.logger = if Ductwork.configuration.logger_source == "rails"
+                            Rails.logger
+                          else
+                            Ductwork::Configuration::DEFAULT_LOGGER
+                          end
+        Ductwork.logger.level = Ductwork.configuration.logger_level
 
         Ductwork::Processes::SupervisorRunner.start!
       end