+++ spec/ductwork/cli_spec.rb
@@ -7,18 +7,18 @@ RSpec.describe Ductwork::CLI do
     logger = instance_double(Logger, :level= => nil)
     config = instance_double(
       Ductwork::Configuration,
-      :logger= => nil,
-      logger: logger,
       logger_level: 0,
       logger_source: "default"
     )
     allow(Ductwork::Processes::SupervisorRunner).to receive(:start!)
     allow(Ductwork::Configuration).to receive(:new).and_return(config)
+    allow(Ductwork).to receive(:logger=).and_call_original
+    allow(Ductwork).to receive(:logger).and_return(logger)
 
     described_class.start!([])
 
     expect(logger).to have_received(:level=).with(0)
-    expect(config).to have_received(:logger=).with(Ductwork::Configuration::DEFAULT_LOGGER)
+    expect(Ductwork).to have_received(:logger=).with(Ductwork::Configuration::DEFAULT_LOGGER)
     expect(Ductwork::Configuration).to have_received(:new)
     expect(Ductwork::Processes::SupervisorRunner).to have_received(:start!)
   end